fix(architecture): split cyclic runtime contracts
