+++ src/infra/approval-native-route-coordinator.ts
@@ -14,8 +14,7 @@ import {
   resolveApprovalRoutedElsewhereNoticeText,
 } from "./approval-native-route-notice.js";
 import { buildChannelApprovalNativeTargetKey } from "./approval-native-target-key.js";
-import type { ApprovalRequestChannelRouteClass } from "./approval-request-account-binding.js";
-import type { ChannelApprovalKind } from "./approval-types.js";
+import type { ApprovalRequestChannelRouteClass, ChannelApprovalKind } from "./approval-types.js";
 import type { ExecApprovalRequest } from "./exec-approvals.js";
 import type { PluginApprovalRequest } from "./plugin-approvals.js";
