+++ src/plugins/gateway-startup-plugin-plan.ts
@@ -37,10 +37,8 @@ import {
   normalizePluginsConfigWithRegistry,
 } from "./plugin-registry-contributions.js";
 import type { PluginRegistrySnapshot } from "./plugin-registry-snapshot.js";
-import {
-  collectConfiguredWorkerProviderIds,
-  normalizeWorkerProviderIds,
-} from "./worker-provider-registry.js";
+import { collectConfiguredWorkerProviderIds } from "./worker-provider-config.js";
+import { normalizeWorkerProviderIds } from "./worker-provider-id.js";
 
 export function resolveChannelPluginIdsFromRegistry(params: {
   manifestRegistry: PluginManifestRegistry;
