+++ src/plugins/plugin-lookup-table.ts
@@ -13,7 +13,7 @@ import {
   type PluginMetadataSnapshot,
 } from "./plugin-metadata-snapshot.js";
 import type { PluginRegistrySnapshot } from "./plugin-registry-snapshot.js";
-import { normalizeWorkerProviderIds } from "./worker-provider-registry.js";
+import { normalizeWorkerProviderIds } from "./worker-provider-id.js";
 
 type PluginLookUpTableMetrics = PluginMetadataSnapshot["metrics"] & {
   startupPlanMs: number;
